add main to functions

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: matmagal <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/22 21:34:58 by matmagal          #+#    #+#             */
-/*   Updated: 2025/04/22 17:08:16 by matmagal         ###   ########.fr       */
+/*   Updated: 2025/04/23 15:04:46 by matmagal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
